Move AstroDataError to factory
